Don't delete things twice

+++ clean.sh
@@ -3,13 +3,17 @@
 # clean sources and buildfiles from all packages
 
 # clean with git
+git clean -dXf
+
 skipped=$(git clean -dXf)
 
 # clean sub repositories
 while read -r line; do
     line=${line//Skipping repository /}
-    echo "Removing repository: $line"
-    rm -rf $line
+    if [ "$line" != "" ]; then
+        echo "Removing repository: $line"
+        rm -rf $line
+    fi
 done <<< "$skipped"
 
 # clean parent directories
